Rename `Runtime` to `Platform`
