+++ examples/structured-output-math.php
@@ -5,7 +5,7 @@ use PhpLlm\LlmChain\Message\Message;
 use PhpLlm\LlmChain\Message\MessageBag;
 use PhpLlm\LlmChain\OpenAI\Model\Gpt;
 use PhpLlm\LlmChain\OpenAI\Model\Gpt\Version;
-use PhpLlm\LlmChain\OpenAI\Runtime\OpenAI;
+use PhpLlm\LlmChain\OpenAI\Platform\OpenAI;
 use PhpLlm\LlmChain\StructuredOutput\ResponseFormatFactory;
 use PhpLlm\LlmChain\StructuredOutput\SchemaFactory;
 use PhpLlm\LlmChain\Tests\StructuredOutput\Data\MathReasoning;
@@ -16,8 +16,8 @@ use Symfony\Component\Serializer\Serializer;
 
 require_once dirname(__DIR__).'/vendor/autoload.php';
 
-$runtime = new OpenAI(HttpClient::create(), getenv('OPENAI_API_KEY'));
-$llm = new Gpt($runtime, Version::gpt4oMini());
+$platform = new OpenAI(HttpClient::create(), getenv('OPENAI_API_KEY'));
+$llm = new Gpt($platform, Version::gpt4oMini());
 $responseFormatFactory = new ResponseFormatFactory(SchemaFactory::create());
 $serializer = new Serializer([new ObjectNormalizer()], [new JsonEncoder()]);
