+++ src/OpenAI/Platform/OpenAI.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace PhpLlm\LlmChain\OpenAI\Runtime;
+namespace PhpLlm\LlmChain\OpenAI\Platform;
 
-use PhpLlm\LlmChain\OpenAI\Runtime;
+use PhpLlm\LlmChain\OpenAI\Platform;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
-final class OpenAI extends AbstractRuntime implements Runtime
+final class OpenAI extends AbstractPlatform implements Platform
 {
     public function __construct(
         private readonly HttpClientInterface $httpClient,
